# application/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "API for managing a mechanic shop, including customers, employees, service tickets, inventory, and services"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions: 
  bearerAuth: # name of the security scheme
    type: apiKey
    name: Authorization
    in: header
    description: "JWT token, prefixed with 'Bearer '"
parameters:
  PageParam:
    name: page
    in: query
    type: integer
    required: false
    default: 1
  LimitParam:
    name: limit
    in: query
    type: integer
    required: false
    default: 10
  SortByParam:
    name: sort_by
    in: query
    type: string
    required: false
  SortOrderParam:
    name: sort_order
    in: query
    type: string
    enum: [asc, desc]
    required: false
...
paths:
...
  /inventory:
    post:
      tags:
        - Inventory
      summary: "Create new inventory item"
      description: "Creates a new inventory item with the provided details. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateInventoryPayload"
      responses:
        201:
          description: "Inventory item created successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Inventory item created successfully"
              data:
                $ref: "#/definitions/InventoryItem"
        400:
          description: "Validation error or duplicate inventory number"
          schema:
            oneOf:
              - $ref: "#/definitions/ValidationErrorResponse"
              - type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "This inventory number already exists."
                  details:
                    type: object
                    properties:
                      inventory_number:
                        type: array
                        items:
                          type: string
                        example: ["This inventory number already exists."]
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden - invalid role"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"

    get:
      tags:
        - Inventory
      summary: "Get inventory items"
      description: "Returns a paginated, filterable list of inventory items. Can filter for deleted or active items."
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/LimitParam"
        - name: sort_by
          in: query
          type: string
          required: false
          enum: [id, inventory_number, name, price, quantity_in_stock]
          description: "Field to sort by. Must be one of the allowed values."
        - $ref: "#/parameters/SortOrderParam"
        - name: deleted
          in: query
          type: boolean
          required: false
          description: "Filter for deleted items (true) or active items (false/not provided)"
        - name: inventory_number
          in: query
          type: string
          required: false
          description: "Filter by inventory number"
        - name: part_name
          in: query
          type: string
          required: false
          description: "Filter by part name"
      responses:
        200:
          description: "Inventory items retrieved successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Operation successful"
              meta:
                $ref: "#/definitions/PaginationMeta"
              data:
                type: array
                items:
                  $ref: "#/definitions/InventoryItem"
        400:
          description: "Invalid sort_by field"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Invalid sort_by field"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /inventory/{inventory_id}:
    get:
      tags:
        - Inventory
      summary: "Get inventory item by ID"
      description: "Returns a single inventory item by its ID"
      parameters:
        - name: inventory_id
          in: path
          required: true
          type: integer
          description: "ID of the inventory item to retrieve"
      responses:
        200:
          description: "Inventory item retrieved successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Operation successful"
              data:
                $ref: "#/definitions/InventoryItem"
        404:
          description: "Inventory not found"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Inventory not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    
    patch:
      tags:
        - Inventory
      summary: "Update inventory item"
      description: "Updates specific fields of an inventory item. Only price and quantity_in_stock can be updated. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: inventory_id
          in: path
          required: true
          type: integer
          description: "ID of the inventory item to update"
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              price:
                type: string
                example: "24.50"
              quantity_in_stock:
                type: integer
                example: 15
      responses:
        200:
          description: "Inventory item updated successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Operation successful"
              data:
                $ref: "#/definitions/InventoryItem"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Inventory not found"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Inventory not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
          
    delete:
      tags:
        - Inventory
      summary: "Delete inventory item"
      description: "Performs a soft delete on the inventory item. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: inventory_id
          in: path
          required: true
          type: integer
          description: "ID of the inventory item to delete"
      responses:
        200:
          description: "Inventory successfully deleted (soft delete)"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Inventory deleted (soft)"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Inventory not found or already deleted"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Inventory already deleted"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /inventory/serialized-parts:
    post:
      tags:
        - Inventory
      summary: "Create a serialized part"
      description: "Creates a new serialized part linked to an inventory item. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateSerializedPartPayload"
      responses:
        201:
          description: "Serialized part created successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Serialized part created successfully"
              data:
                $ref: "#/definitions/SerializedPartItem"
        400:
          description: "Validation error or duplicate serial number"
          schema:
            oneOf:
              - $ref: "#/definitions/ValidationErrorResponse"
              - type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  message:
                    type: string
                    example: "This serialized part already exists"
                  details:
                    type: object
                    properties:
                      serial_number:
                        type: array
                        items:
                          type: string
                        example: ["This serialized part already exists."]
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden - invalid role"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
          
    get:
      tags:
        - Inventory
      summary: "Get all serialized parts"
      description: "Returns a paginated list of serialized parts with their related inventory details."
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/LimitParam"
        - $ref: "#/parameters/SortByParam"
        - $ref: "#/parameters/SortOrderParam"
        - name: status
          in: query
          type: string
          enum: [available, used, defective]
          required: false
          description: "Filter by part status"
      responses:
        200:
          description: "Serialized parts retrieved successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Operation successful"
              meta:
                $ref: "#/definitions/PaginationMeta"
              data:
                type: array
                items:
                  $ref: "#/definitions/SerializedPartItem"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
  /inventory/serialized-parts/{part_id}:
    get:
      tags:
        - Inventory
      summary: "Get serialized part by ID"
      description: "Returns a single serialized part by its ID"
      parameters:
        - name: part_id
          in: path
          required: true
          type: integer
          description: "ID of the serialized part to retrieve"
      responses:
        200:
          description: "Serialized part retrieved successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Operation successful"
              data:
                $ref: "#/definitions/SerializedPartItem"
        404:
          description: "Serialized part not found"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Serialized part not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
            
    patch:
      tags:
        - Inventory
      summary: "Update serialized part status"
      description: "Updates the status of a serialized part. Only the status field can be updated. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: part_id
          in: path
          required: true
          type: integer
          description: "ID of the serialized part to update"
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
                enum: [available, used, defective]
                example: "defective"
      responses:
        200:
          description: "Serialized part updated successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Operation successful"
              data:
                $ref: "#/definitions/SerializedPartItem"
        400:
          description: "Invalid status value"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Invalid status value"
              details:
                type: object
                properties:
                  status:
                    type: array
                    items:
                      type: string
                    example: ["Invalid status value."]
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Serialized part not found"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Serialized part not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
    
    delete:
      tags:
        - Inventory
      summary: "Delete serialized part"
      description: "Performs a soft delete on the serialized part. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: part_id
          in: path
          required: true
          type: integer
          description: "ID of the serialized part to delete"
      responses:
        200:
          description: "Serialized part deleted successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Serialized part deleted successfully"
        400:
          description: "Serialized part already deleted"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Serialized part is already deleted"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Serialized part not found"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Serialized part not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"

#MARK: service paths
...
definitions:
...
  ValidationErrorResponse:
    type: object
    properties:
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Invalid input"
      errors:
        type: object
        properties:
          field_name:
            type: array
            items:
              type: string
            example: ["Missing data for required field."]
    examples:
      application/json:
        status: "error"
        message: "Invalid input"
  UnauthorizedResponse:
    type: object
    properties:
      message:
        type: string
        example: "Authorization header is required"
    examples:
      application/json:
        message: "Authorization header is required"
  ForbiddenResponse:
    type: object
    properties:
      message:
        type: string
        example: "Forbidden - invalid role"
    examples:
      application/json:
        message: "Forbidden - invalid role"
...
  ServerErrorResponse:
    type: object
    properties:
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Internal server error"
    examples:
      application/json:
        status: "error"
        message: "Internal server error"
...
  PaginationMeta:
    type: object
    properties:
      pagination:
        type: object
        properties:
          has_next:
            type: boolean
            example: false
          has_prev:
            type: boolean
            example: false
          limit:
            type: integer
            example: 10
          page:
            type: integer
            example: 1
          total_items:
            type: integer
            example: 1
          total_pages:
            type: integer
            example: 1
...
  CreateInventoryPayload:
    type: object
    required:
      - name
      - inventory_number
      - price
      - quantity_in_stock
    properties:
      name:
        type: string
        example: "Wheel"
      inventory_number:
        type: string
        example: "SP-201"
        description: "Must be unique"
      price:
        type: string
        example: "14.99"
      desc:
        type: string
        example: "some car wheel"
      quantity_in_stock:
        type: integer
        example: 50
  InventoryItem:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: "Oil Filter"
      inventory_number:
        type: string
        example: "FLT-2203"
      price:
        type: string
        example: "24.50"
      desc:
        type: string
        example: "Standard replacement oil filter compatible with most passenger vehicles"
      quantity_in_stock:
        type: integer
        example: 15
      is_deleted:
        type: boolean
        example: false
  CreateSerializedPartPayload:
    type: object
    required:
      - serial_number
      - status
      - inventory_id
    properties:
      serial_number:
        type: string
        example: "TZ-7824"
      status:
        type: string
        enum: [available, used, defective]
        example: "available"
      inventory_id:
        type: integer
        example: 5
  SerializedPartItem:
    type: object
    properties:
      id:
        type: integer
        example: 5
      serial_number:
        type: string
        example: "TZ-7824"
      inventory_id:
        type: integer
        example: 5
      status:
        type: string
        enum: [available, used, defective]
        example: "available"
      is_deleted:
        type: boolean
        example: false
      inventory:
        type: object
        properties:
          id:
            type: integer
            example: 5
          name:
            type: string
            example: "Spark Plug"
          inventory_number:
            type: string
            example: "SP-200"
          price:
            type: string
            example: "14.99"
          desc:
            type: string
            example: "Spark plug forengines"

# MARK: Service Definitions
...